requirements.txt
torch>=2.0.0
torchvision>=0.15.0
scipy>=1.10.0
numpy>=1.24.0
Pillow>=9.0.0
tqdm>=4.64.0
pyyaml>=6.0
einops>=0.6.0
pytest>=7.0.0
